split huge roberta file into structured project
