+++ main.py
@@ -1,0 +1,168 @@
+"""The main file with the training and evaluation loop."""
+import torch
+import torch.optim as optim
+from torch.nn import CrossEntropyLoss, MSELoss
+import numpy as np
+import scipy
+from sklearn.metrics import matthews_corrcoef, f1_score, accuracy_score
+from transformers import RobertaTokenizer
+from tqdm import tqdm
+from torch.utils.data import DataLoader
+from data.glue import GLUEDataset
+from models.lora_roberta import LoraWrapperRoberta
+
+
+def compute_metrics(task_name, preds, labels):
+    """Compute metrics for GLUE tasks"""
+    if task_name == 'cola':
+        return {'mcc': matthews_corrcoef(labels, preds)}
+    elif task_name in ['sst2', 'mrpc', 'qqp', 'qnli', 'rte', 'wnli']:
+        acc = accuracy_score(labels, preds)
+        f1 = f1_score(labels, preds)
+        return {'accuracy': acc, 'f1': f1}
+    elif task_name == 'stsb':
+        return {'pearson': np.corrcoef(preds, labels)[0, 1],
+                'spearman': scipy.stats.spearmanr(preds, labels)[0]}
+    elif task_name == 'mnli':
+        return {'accuracy': accuracy_score(labels, preds)}
+
+
+def train_epoch(model, train_loader, optimizer, criterion, device, task_name):
+    model.train()
+    total_loss = 0
+    all_preds = []
+    all_labels = []
+
+    progress_bar = tqdm(train_loader, desc='Training')
+    for batch in progress_bar:
+        input_ids = batch['input_ids'].to(device)
+        attention_mask = batch['attention_mask'].to(device)
+        labels = batch['label'].to(device)
+
+        optimizer.zero_grad()
+        outputs = model(input_ids, attention_mask)
+
+        if task_name == 'stsb':
+            loss = criterion(outputs.squeeze(), labels)
+            preds = outputs.squeeze().detach().cpu().numpy()
+        else:
+            loss = criterion(outputs, labels)
+            preds = torch.argmax(outputs, dim=1).cpu().numpy()
+
+        loss.backward()
+        optimizer.step()
+
+        total_loss += loss.item()
+        all_preds.extend(preds)
+        all_labels.extend(labels.cpu().numpy())
+
+        progress_bar.set_postfix({'loss': loss.item()})
+
+    metrics = compute_metrics(task_name, np.array(all_preds), np.array(all_labels))
+    metrics['loss'] = total_loss / len(train_loader)
+    return metrics
+
+
+def evaluate(model, eval_loader, criterion, device, task_name):
+    model.eval()
+    total_loss = 0
+    all_preds = []
+    all_labels = []
+
+    with torch.no_grad():
+        for batch in tqdm(eval_loader, desc='Evaluating'):
+            input_ids = batch['input_ids'].to(device)
+            attention_mask = batch['attention_mask'].to(device)
+            labels = batch['label'].to(device)
+
+            outputs = model(input_ids, attention_mask)
+
+            if task_name == 'stsb':
+                loss = criterion(outputs.squeeze(), labels)
+                preds = outputs.squeeze().cpu().numpy()
+            else:
+                loss = criterion(outputs, labels)
+                preds = torch.argmax(outputs, dim=1).cpu().numpy()
+
+            total_loss += loss.item()
+            all_preds.extend(preds)
+            all_labels.extend(labels.cpu().numpy())
+
+    metrics = compute_metrics(task_name, np.array(all_preds), np.array(all_labels))
+    metrics['loss'] = total_loss / len(eval_loader)
+    return metrics
+
+
+def main():
+    # Configuration
+    task_name = 'sst2'  # Choose from GLUE tasks
+    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+    batch_size = 32
+    num_epochs = 3
+    learning_rate = 1e-4
+    max_length = 128
+
+    print(f"Running GLUE task: {task_name}")
+    print(f"Using device: {device}")
+
+    # Initialize tokenizer
+    tokenizer = RobertaTokenizer.from_pretrained('roberta-base')
+
+    # Load datasets
+    train_dataset = GLUEDataset(task_name, 'train', tokenizer, max_length)
+    eval_dataset = GLUEDataset(task_name, 'validation', tokenizer, max_length)
+
+    train_loader = DataLoader(train_dataset, batch_size=batch_size, shuffle=True)
+    eval_loader = DataLoader(eval_dataset, batch_size=batch_size)
+
+    # Initialize model
+    num_labels = train_dataset.num_labels
+    model = LoraWrapperRoberta(
+        task_type='glue',
+        num_classes=num_labels,
+        lora_rank=8,
+        train_biases=True,
+        train_embedding=False,
+        train_layer_norms=True
+    )
+    model.to(device)
+
+    # Print parameter counts
+    total_params = sum(p.numel() for p in model.parameters())
+    trainable_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
+    print(f"Total parameters: {total_params:,}")
+    print(f"Trainable parameters: {trainable_params:,}")
+    print(f"Percentage of trainable parameters: {trainable_params / total_params * 100:.2f}%")
+
+    # Initialize optimizer and loss function
+    optimizer = optim.AdamW(model.parameters(), lr=learning_rate)
+    criterion = MSELoss() if task_name == 'stsb' else CrossEntropyLoss()
+
+    # Training loop
+    best_metric = float('-inf')
+    for epoch in range(num_epochs):
+        print(f"\nEpoch {epoch + 1}/{num_epochs}")
+
+        # Training
+        train_metrics = train_epoch(model, train_loader, optimizer, criterion, device, task_name)
+        print("\nTraining metrics:")
+        for k, v in train_metrics.items():
+            print(f"{k}: {v:.4f}")
+
+        # Evaluation
+        eval_metrics = evaluate(model, eval_loader, criterion, device, task_name)
+        print("\nValidation metrics:")
+        for k, v in eval_metrics.items():
+            print(f"{k}: {v:.4f}")
+
+        # Save best model
+        current_metric = eval_metrics['accuracy'] if 'accuracy' in eval_metrics else eval_metrics['mcc']
+        if current_metric > best_metric:
+            best_metric = current_metric
+            model.save_lora_state_dict(f'lora_weights_{task_name}_best.pt')
+
+    print("\nTraining completed!")
+    print(f"Best validation metric: {best_metric:.4f}")
+
+if __name__ == '__main__':
+    main()
